Add files via upload
Attached you will finr final project for ITBootcamp Java fundamentials cours.
For this project we got Dictionary database, and file Book.
The task was to process the book and dictionary. The program should find and save all the words from the book that are not in the dictionary. Dictionary data base should be expanded with new table that will store new words which are not in initial database.
For each word in dictionary find how often it apperas in book.
Find 20 most repeated words.
Create file with all the known, and new words, and they should be ordered in lexicographic order.